(PC-19593) refactor(navigation): move RootNavigator tests outside __tests__ folder
