+++ src/features/navigation/RootNavigator/RootNavigator.web.test.tsx
@@ -3,14 +3,18 @@ import React from 'react'
 import { mocked } from 'ts-jest/utils'
 
 import { useMustUpdateApp } from 'features/forceUpdate/useMustUpdateApp'
+import { useCurrentRoute } from 'features/navigation/helpers'
 import { useSplashScreenContext } from 'libs/splashscreen'
-import { render, flushAllPromisesWithAct } from 'tests/utils'
+import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
+import { flushAllPromisesWithAct, render } from 'tests/utils/web'
 
-import { RootNavigator } from '../RootNavigator'
+import { RootNavigator } from './RootNavigator'
 
 const mockUseSplashScreenContext = mocked(useSplashScreenContext)
 const mockedUseMustUpdateApp = mocked(useMustUpdateApp)
+const mockUseCurrentRoute = mocked(useCurrentRoute)
 
+jest.mock('features/cookies/helpers/useIsCookiesListUpToDate')
 jest.mock('features/forceUpdate/useMustUpdateApp')
 jest.mock('@react-navigation/native', () => jest.requireActual('@react-navigation/native'))
 jest.mock('features/auth/AuthContext')
@@ -23,16 +27,15 @@ jest.mock('features/navigation/TabBar/TabNavigator', () => ({
 jest.mock('features/navigation/RootNavigator/useInitialScreenConfig', () => ({
   useInitialScreen: () => 'TabNavigator',
 }))
-jest.mock('features/navigation/helpers', () => ({
-  useCurrentRoute: () => ({ name: 'TabNavigator', key: 'key' }),
-}))
+jest.mock('features/navigation/helpers')
 jest.mock('libs/splashscreen')
 
 describe('<RootNavigator />', () => {
-  beforeEach(() => mockedUseMustUpdateApp.mockReturnValue(true))
+  beforeEach(() => {
+    mockUseCurrentRoute.mockReturnValue({ name: 'TabNavigator', key: 'key' })
+  })
 
   it('should NOT display PrivacyPolicy if splash screen is not yet hidden', async () => {
-    mockedUseMustUpdateApp.mockReturnValueOnce(false)
     mockUseSplashScreenContext.mockReturnValueOnce({ isSplashScreenHidden: false })
     const renderAPI = await renderRootNavigator()
 
@@ -41,17 +44,26 @@ describe('<RootNavigator />', () => {
   })
 
   it('should display PrivacyPolicy if splash screen is hidden', async () => {
-    mockedUseMustUpdateApp.mockReturnValueOnce(false)
     mockUseSplashScreenContext.mockReturnValueOnce({ isSplashScreenHidden: true })
 
     const renderAPI = await renderRootNavigator()
 
-    const privacyPolicyTitle = renderAPI.queryByText('Respect de ta vie privée')
+    const privacyPolicyTitle = renderAPI.getByText('Respect de ta vie privée')
     expect(privacyPolicyTitle).toBeTruthy()
   })
 
-  it('should not display quick access button in native', async () => {
+  it('should display quick access button if show tabBar and current route is TabNavigator', async () => {
     mockUseSplashScreenContext.mockReturnValueOnce({ isSplashScreenHidden: true })
+
+    const renderAPI = await renderRootNavigator()
+
+    const quickAccessButton = renderAPI.queryByText('Accéder au menu de navigation')
+    expect(quickAccessButton).toBeTruthy()
+  })
+
+  it('should not display quick access button if current route is not TabNavigator', async () => {
+    mockUseSplashScreenContext.mockReturnValueOnce({ isSplashScreenHidden: true })
+    mockUseCurrentRoute.mockReturnValueOnce({ name: 'Offer', key: 'key' })
 
     const renderAPI = await renderRootNavigator()
 
@@ -62,17 +74,23 @@ describe('<RootNavigator />', () => {
 
 describe('ForceUpdate display logic', () => {
   it('should display force update page when global variable is set', async () => {
+    mockUseCurrentRoute.mockReturnValueOnce({ name: 'TabNavigator', key: 'key' })
+    mockedUseMustUpdateApp.mockReturnValueOnce(true)
+
     const rootNavigator = await renderRootNavigator()
 
-    expect(rootNavigator.queryAllByText('Mise à jour de l’application')).toBeTruthy()
+    expect(rootNavigator.queryAllByText("Mise à jour de l'application")).toBeTruthy()
   })
 })
 
 async function renderRootNavigator() {
   const renderAPI = render(
-    <NavigationContainer>
-      <RootNavigator />
-    </NavigationContainer>
+    // eslint-disable-next-line local-rules/no-react-query-provider-hoc
+    reactQueryProviderHOC(
+      <NavigationContainer>
+        <RootNavigator />
+      </NavigationContainer>
+    )
   )
 
   await flushAllPromisesWithAct()
